step 1: pdbDownloader -> add the capability to detect either the file names are a list or a csv. then download. instead of having it to convert in gui part
step 2:
cleanup change seperator to make sense

step 3:
delete the csv builder and have it to be created with the functions when we want to download the files based on info we want (different than the downloader that works with lists)

step 4:
figure something out for the deleter part.

step 5:
in the gui code, remove the create csv, remove the update csv with pdb.

step 6:
make the gui code to call to the download of pdb files based on the functions that is checked and added to a list of the function names